# openapi.yaml
openapi: 3.0.3
info:
  title: rebox
  description: |
    **rebox** loads a **public URL** (headless Chromium) and returns **text**, **image**, or **YouTube transcript**.

    **Plain URL (recommended):** `GET|POST /rebox/text`, `/rebox/image`, `/rebox/audio` — pass the full URL as **`?url=`** (encode the value once if it contains `?`, `&`, `#`) or as **`{ "url": "https://..." }`** in the POST JSON body (exact string, no path encoding). Options in query and/or JSON body on POST.

    **Legacy path segment:** `/rebox/<encodeURIComponent(url)>/text` (slashes in the URL become `%2F` inside one path segment).

    **Example:** `POST /rebox/text` with body `{"url":"https://learn.ai-native.pro/","timeout_ms":90000}`

    Defaults: `wait_until=domcontentloaded`, `timeout_ms=60000`, **`settle_ms`** from env `REBOX_DEFAULT_SETTLE_MS` (default **2000** ms).

    Image body is raw PNG/WebP bytes; optional query `save_as`, `filename_source`, `fullPage`, `scroll_full_page`, etc. Full-page shots default to scrolling the page (up to **`REBOX_FULL_PAGE_SCROLL_MAX_MS`**, default **25000** ms) so lazy blocks (e.g. Habr comments) expand before capture.

    When `REBOX_API_KEYS` is set: send **`X-API-Key`** or **`Authorization: Bearer`** token (also **`Authorization: ApiKey`**). Same routes accept **GET** or **POST**; on POST, optional JSON body overrides query for duplicate keys.

    SSRF checks on every target URL.
  version: 0.5.2
servers:
  - url: /
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /rebox/text:
    get:
      tags: [rebox]
      summary: Text — full URL in query `url` (no path encoding)
      operationId: reboxTextPlainGet
      parameters:
        - $ref: '#/components/parameters/PlainTargetUrl'
        - $ref: '#/components/parameters/WaitUntil'
        - $ref: '#/components/parameters/TimeoutMs'
        - $ref: '#/components/parameters/SettleMs'
        - $ref: '#/components/parameters/ViewportWidth'
        - $ref: '#/components/parameters/ViewportHeight'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/UserAgent'
        - name: markdown
          in: query
          schema: { type: boolean, default: true }
        - name: separateMarkdown
          in: query
          schema: { type: boolean, default: false }
        - name: includeRawHtml
          in: query
          schema: { type: boolean, default: false }
        - name: maxRawHtmlChars
          in: query
          schema: { type: integer }
        - name: content_selector
          in: query
          schema: { type: string }
        - name: debug
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: JSON with `visibleText` and Defuddle `article`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxTextResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/Timeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
    post:
      tags: [rebox]
      summary: 'POST JSON with required url (full https string) and optional fields; merges with query'
      operationId: reboxTextPlainPost
      parameters:
        - $ref: '#/components/parameters/WaitUntil'
        - $ref: '#/components/parameters/TimeoutMs'
        - $ref: '#/components/parameters/SettleMs'
        - $ref: '#/components/parameters/ViewportWidth'
        - $ref: '#/components/parameters/ViewportHeight'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/UserAgent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              additionalProperties: true
              properties:
                url:
                  type: string
                  description: Full target URL as a single string
      responses:
        '200':
          description: JSON with `visibleText` and Defuddle `article`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxTextResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/Timeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
  /rebox/image:
    get:
      tags: [rebox]
      summary: Image — full URL in query `url`
      operationId: reboxImagePlainGet
      parameters:
        - $ref: '#/components/parameters/PlainTargetUrl'
        - $ref: '#/components/parameters/WaitUntil'
        - $ref: '#/components/parameters/TimeoutMs'
        - $ref: '#/components/parameters/SettleMs'
        - $ref: '#/components/parameters/ViewportWidth'
        - $ref: '#/components/parameters/ViewportHeight'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/UserAgent'
        - name: fullPage
          in: query
          schema: { type: boolean, default: true }
        - name: scroll_full_page
          in: query
          description: Scroll to load lazy content before a full-page capture (default true). Ignored when fullPage is false.
          schema: { type: boolean, default: true }
        - name: format
          in: query
          schema:
            type: string
            enum: [png, webp]
            default: png
        - name: maxHeightPx
          in: query
          schema: { type: integer }
        - name: quality
          in: query
          schema: { type: integer, minimum: 1, maximum: 100 }
        - name: save_as
          in: query
          schema: { type: string, maxLength: 200 }
        - name: filename_source
          in: query
          schema:
            type: string
            enum: [host, title]
            default: host
      responses:
        '200':
          description: Raw image
          content:
            image/png:
              schema: { type: string, format: binary }
            image/webp:
              schema: { type: string, format: binary }
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/Timeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
    post:
      tags: [rebox]
      summary: Image — URL in JSON body
      operationId: reboxImagePlainPost
      parameters:
        - $ref: '#/components/parameters/WaitUntil'
        - $ref: '#/components/parameters/TimeoutMs'
        - $ref: '#/components/parameters/SettleMs'
        - $ref: '#/components/parameters/ViewportWidth'
        - $ref: '#/components/parameters/ViewportHeight'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/UserAgent'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              additionalProperties: true
              properties:
                url: { type: string }
                fullPage: { type: boolean }
                scroll_full_page: { type: boolean }
                format: { type: string, enum: [png, webp] }
      responses:
        '200':
          description: Raw image
          content:
            image/png:
              schema: { type: string, format: binary }
            image/webp:
              schema: { type: string, format: binary }
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/Timeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
  /rebox/audio:
    get:
      tags: [rebox]
      summary: YouTube audio — full URL in query `url`
      operationId: reboxAudioPlainGet
      parameters:
        - $ref: '#/components/parameters/PlainTargetUrl'
        - name: lang
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Transcript segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxAudioResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not YouTube / no transcript
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
    post:
      tags: [rebox]
      summary: YouTube audio — URL in JSON body
      operationId: reboxAudioPlainPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              additionalProperties: true
              properties:
                url: { type: string }
                lang: { type: string }
      responses:
        '200':
          description: Transcript segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxAudioResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Not YouTube / no transcript
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
  /rebox/{encodedUrl}/text:
    get:
      tags: [rebox]
      summary: Render URL, extract article + visible plain text
      operationId: reboxText
      parameters:
        - $ref: '#/components/parameters/EncodedUrl'
        - $ref: '#/components/parameters/WaitUntil'
        - $ref: '#/components/parameters/TimeoutMs'
        - $ref: '#/components/parameters/SettleMs'
        - $ref: '#/components/parameters/ViewportWidth'
        - $ref: '#/components/parameters/ViewportHeight'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/UserAgent'
        - name: markdown
          in: query
          schema: { type: boolean, default: true }
        - name: separateMarkdown
          in: query
          schema: { type: boolean, default: false }
        - name: includeRawHtml
          in: query
          schema: { type: boolean, default: false }
        - name: maxRawHtmlChars
          in: query
          schema: { type: integer }
        - name: content_selector
          in: query
          schema: { type: string }
        - name: debug
          in: query
          schema: { type: boolean, default: false }
      responses:
        '200':
          description: JSON with `visibleText` (DOM) and Defuddle `article`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxTextResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/Timeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
    post:
      tags: [rebox]
      summary: Same as GET — use POST; options in query and/or JSON body (body wins on duplicate keys)
      operationId: reboxTextPost
      parameters:
        - $ref: '#/components/parameters/EncodedUrl'
        - $ref: '#/components/parameters/WaitUntil'
        - $ref: '#/components/parameters/TimeoutMs'
        - $ref: '#/components/parameters/SettleMs'
        - $ref: '#/components/parameters/ViewportWidth'
        - $ref: '#/components/parameters/ViewportHeight'
        - $ref: '#/components/parameters/Locale'
        - $ref: '#/components/parameters/UserAgent'
        - name: markdown
          in: query
          schema: { type: boolean, default: true }
        - name: separateMarkdown
          in: query
          schema: { type: boolean, default: false }
        - name: includeRawHtml
          in: query
          schema: { type: boolean, default: false }
        - name: maxRawHtmlChars
          in: query
          schema: { type: integer }
        - name: content_selector
          in: query
          schema: { type: string }
        - name: debug
          in: query
          schema: { type: boolean, default: false }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: Optional fields mirroring query (e.g. timeout_ms, settle_ms)
      responses:
        '200':
          description: JSON with `visibleText` (DOM) and Defuddle `article`
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxTextResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '408':
          $ref: '#/components/responses/Timeout'
        '413':
          $ref: '#/components/responses/PayloadTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
...
  /rebox/{encodedUrl}/audio:
    get:
      tags: [rebox]
      summary: YouTube transcript (non-YouTube URL → 404)
      operationId: reboxAudio
      parameters:
        - $ref: '#/components/parameters/EncodedUrl'
        - name: lang
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Transcript segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxAudioResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No transcript / not YouTube
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
    post:
      tags: [rebox]
      summary: Same as GET — YouTube transcript JSON; optional JSON body for lang etc.
      operationId: reboxAudioPost
      parameters:
        - $ref: '#/components/parameters/EncodedUrl'
        - name: lang
          in: query
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Transcript segments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReboxAudioResponse'
        '304':
          description: Not modified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: No transcript / not YouTube
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '502':
          $ref: '#/components/responses/BadGateway'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    PlainTargetUrl:
      name: url
      in: query
      required: true
      description: |
        Full target URL as one string. If it contains `?`, `&`, or `#`, encode the **value** with `encodeURIComponent` once (standard query encoding).
      schema:
        type: string
...
    WaitUntil:
      name: wait_until
      in: query
      schema:
        type: string
        enum: [load, domcontentloaded, networkidle]
        default: domcontentloaded
    TimeoutMs:
      name: timeout_ms
      in: query
      schema:
        type: integer
        default: 60000
    SettleMs:
      name: settle_ms
      in: query
      description: Extra ms to wait after navigation (hydration). Overrides `REBOX_DEFAULT_SETTLE_MS` when set.
      schema:
        type: integer
        minimum: 0
        maximum: 30000
    ViewportWidth:
      name: viewport_w
      in: query
      schema:
        type: integer
        default: 1280
    ViewportHeight:
      name: viewport_h
      in: query
      schema:
        type: integer
        default: 720
    Locale:
      name: locale
      in: query
      schema: { type: string }
    UserAgent:
      name: user_agent
      in: query
      schema: { type: string }
  responses:
    BadRequest:
      description: Bad request / SSRF
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Timeout:
      description: Timeout
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    PayloadTooLarge:
      description: Too large
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    BadGateway:
      description: Upstream / browser error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
  schemas:
...
    ReboxTextResponse:
      type: object
      required: [finalUrl, article, visibleText, timingsMs]
      properties:
        status: { type: string, enum: [success] }
        finalUrl: { type: string, format: uri }
        article:
          type: object
          additionalProperties: true
        visibleText:
          type: string
          description: Live DOM innerText (SPA-friendly)
        rawHtml: { type: string }
        timingsMs:
          type: object
          properties:
            navigation: { type: number }
            defuddle: { type: number }
        requestId: { type: string }
...